+++ lib/http/rest-client.js
@@ -3,7 +3,7 @@ import humps from 'humps';
 import qs from 'qs'; // TODO: Check with Alexandar if httpAgent is supposed to be added
 
 import RestError from './rest-error';
-import { RES_STATUS_DELAYED, RES_STATUS_ACCEPTED } from './../util/constants';
+import { RES_STATUS_DELAYED, RES_STATUS_ACCEPTED } from './../constants';
 
 class RestClient {
   /**
