+++ lib/util/url.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import URI from 'urijs';
-import { BASE_URL, URL_PROTOCOL } from './constants';
+import { BASE_URL, URL_PROTOCOL } from './../constants';
 
 /**
  * Get base url of the API
