+++ lib/util/version.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import { VERSIONS } from './constants';
+import { VERSIONS } from './../constants';
 
 /**
  * Returns Version from VersionMap
